LPS-120270 Move to a different package
the registry does not need to be exported
