+++ portal-impl/src/com/liferay/portal/security/auth/packageinfo
@@ -1,1 +1,1 @@
-version 3.0.0
+version 2.0.0
